BAU: Refactor some of the x509 references
A little bit of optimization by creating an instance variable representing the
x509 object in the validator instead of the method
and executing the conversion all the time.
Also added the x509 property on to the Certificate class to promote separation of concers
and not use the method from the validator.
This will also help us exposing other cert details going forward.