Add vsp column to SpComponent
This will be used to discern whether an SP component is a Verify one or
not. By moving to the separate component classes it will be easier to
reference the correct component if we do not have a string designating a
VSP.

Create SpComponent and MsaComponent constants and use these in the
relevant component classes.