Correctly validate uniqueness of entity id
The service entity id needs to be unique. This should be
done in the serivce and not in the service event as `events`
themselves have no concept of entity ids.

Merged in the service errors into the service event errors just to make
it easier to represent them in the views.