Correctly publish metadata
We had previously made an incorrect assumption about the publishing of
the metadata. It should always be a the latest metadata for all
components regardless of type.

The component_spec should now show a complete example of the published
metadata structure.

Also removed some tests from component_spec which were already being
tested in other specs.