Fix showing of certificates for specific component types
Previously when visiting a component page which happened to share the
same ID of a component of a different type, the certificates relating to
the alternative component would also be shown on the page. This was
because the has_many association did not have knowledge of the component
type and therefore would simply return all the certificates which had a
component id that matched.

One option would have been to use the same approach that was done for
the services which is create separate columns for each type of
component. We decided not to do this as we wanted to maintain the
concept that a certificate can only be associated with a single
component regardless of type.

has_many association on the component allows for the intended component
type to be accessed.