Adjust previous migrations
We need the environment column to have a default value or the migrations
will fail. Changing the previous migration files is awkward but should
not be a problem for local machines that have already run them as they
will be skipped.

The Self Service app has yet to apply these migrations so should run
them as intended.