Remove secrets and env specific files
The .env.test and .env.development files were only added when trying to
debug the deployment of the self service app.

This PR removes the secrets file and the environment specific .env
files. The config is moved into the regular <environment>.rb files. This
allows us to use `.fetch` in production where we want the app to throw
and exception if certain config is not set in the environment.

Developers are still able to use their own .env file locally to override
any of these variables when required.