Remove environment constants
This could be a little contentious. In response to [this comment](https://github.com/alphagov/verify-self-service/pull/125#discussion_r316805353)
this PR removes all the environment constants. The idea behind this is
that they should already be defined in the `hub_environments`.

This prompted a little refactor in order to make use of Factory Bot a
little more. The upload_certificate_event still needs to be a factory at
some point. Perhaps another PR . . . .