MdAKXfZX: make the user available in the tests
To be able to test components that belong to a team, the component team_id needs to be
set to the user team_id. And to do that the user needs to be made visible in the tests.
The team_id is a SecureRandom.uuid so passing the user and pulling it out seems like the easiest
worko around.