Support the uploading of certificate files
This adds the ability to upload files for the user journey part of
the Self Service app.

So far it only validates the content type of the files and passes
the contents back to the controller. There will need to be some
changes in a future PR with how validations are done in order to
actually validate the certificate whether it has been pasted in or
extracted from a file.

For the content types we are allowing:

application/x-x509-ca-cert => .crt and .der
application/x-pem-file => .pem

We disallow everything else including application/pkcs8 which is the
content type for private keys.