Validate certificates in user journey flow
We need to be able to validate certificates separate from the event
models. As the user moves between pages in the user journey we need to
be able to validate the certificate twice; once when they upload/paste a
certificate, and then again after they have viewed it and submitted it.

We do not want to trigger validation on the Certificate model itself as
this would mean that validation would occur up to four times while they
progress through the journey, as opposed to just to the two we require.