BAU: Allow users to leave the sign-in/up flow
Currently, when user enters their password correctly, their session gets populated
and can only finish or fail the journey, no option to cancel.
Even if they visit the homepage again, the session still remembers their details so will be offering them
the TOTP prompt (or whatever step in the journey they were). The user needs to either fail
or delete their cookies in order to get out of the flow.
This adds a cancel button to all the sign-in/up flow screens to reset their session.

Also, added test and localized the strings in there.