Add an integrity checker
A fail-safe mechanism to make sure groups in Cognito align with teams in the app.
It could happen that a Cognito group exists without a corresponding team, meaning we
are unable to manage it or see.
On the startup it will compare the cognito groups with teams and create any missing teams.
Also, I have set the puma workers number to one, since it traffic is not expected to be high and also it
would cause the checker to run twice, potentially creating race conditions.