A successful Rails 6.0 upgrade
OK some caveats...  I have turned off sentry in development/test.  I
have done this because it was masking errors locally as it was uanble
to connect to an instance of sentry.

Rspec 3.8 is not compatible with Rails 6 and technically Rails 4.0
is not out yet.  So to make Rspec work I have had to upgrade the Gem
to the beta version.  I think this will stop us upgrading to Rails 6.0
at the moment.  I haven't been able to find an expected release date
for Rspec 4.0 so its a choice to having a beta package in our app
(even if it is for testing) or to stick with Rails 5.2.3.