R4q1NXt5: Add page titles
We had the page title static and fixed thorough the application.
This change makes the page title dynamic based on the view. I had to
retrospectively add the page titles to the views. This is particularly important
for any web analytics and for the users.

Also, I've added a check to raise an error if someone forgets to add a page title.