hNmxRDdx: Correct the DEPLOYING status on signing certs
Until now the status was driven by the number of enabled signing certs on the component.
This PR adds the real logic with a 10 minute period for which it will
display the deploying status and all the corresponding messages on
primary/secondary certs. I also reshuffled the logic slightly to be more readable.

In the follow-up PR I'll do the encryption cert logic and the TODO bit.