1NDRKTxw: Support for disabling certs
Currently users are not able to disable ("Stop using") the secondary/old certs.

This is to implement the logic to allow them to disable the old one when dual running.

I've also updated the validation on the change event to prevent disable a cert
if it's the only one and the component would end up without certs.