BUGFIX: Do not validate twice on disable/enable event
Currently, when we disable a cert, we get a confusing error message saying
the component has to have at least one certificate.
This is due to the fact, that validation happens during the event creation and
then we call `valid?` again, which now adds the error because
the cert was disabled during the event itself.
This change looks at the errors on the model instead of re-running the validation.