Create CertStatusUpdater module
This module provides code which handles checking Hub for the current status of a particular certificate, and taking appropriate action based on the result.  If the given cert is included in the list of certs returned by Hub, then the in_use_at field on that cert will be set to a timestamp representing the moment by which we expect all Hub cert caches to have expired, and thus can trust that the cert is being used throughout all of Hub.