BAU: Show more details about certificates for admins
The admin interface was quite limited in what it was showing about the certificates attached
to a component. This changes exposes the common name, expiry date and the certificate value.
This should make it easier to see what certs exactly are in use.

Also, simplified the partial views to use only one and removed a test as it was
no longer possible to test an invalid test. However, this logic is covered within the
event validation and should not be tested via system tests anyway.