BAU: Fix the incorrect reference
The heading of the table was referring to `user_journey.certificate` however,
it's defined twice, once as a value and second time as a group including sub-keys.
The page was rendering the sub-keys. It wasn't visible to the user because of
being it visually hidden for screen readers.

This PR changes the certificate key to certificate_heading to remove the duplication.