gbOCPRmz: Make usernames/emails case insensitive
Currently, the emails used as usernames are case sensitive. This adds a downcase method on
models when we're handling emails.
I did consider putting to the API calls cognito only, but we often store the email
in the session for further processing (MFA enrolment) which would create
discrepencies.
Open to other suggestions.