Email templates for when cert is ready
When a certificate that has been added by a reying party is confirmed to be in use by the Hub, we will send them an email to confirm that the cert is ready.  This commit adds code which chooses which template to use and sends the email, but at present nothing calls it.

https://trello.com/c/OjUWN78Y/782-schedule-an-email-when-cert-is-confirmed-to-be-in-use-by-hub