BAU: Fix failing healthchecks
We can't currently deploy as the healthchecks are failing. This is due to a change in data
structure of the hub_environments. This aligns the dev, test and production
format to be the same to avoid this in future.

We were also missing the authentication_header config on test.rb