RK8WfX73: Fix the logic for finding expiring certs
The current logic for finding expiring certs is flawed. It works under the assumption
that certs are in use when their 'enabled' attribute is set to true.
It's the case for signing certs. However, encryption certs are behaving differently.
Encryption certs are replaced since there is one-to-one relationship
from a component perspective.

This change adds a new method on the component to check whether a given cert
is active/used by the component. This method is then used to find the certs which
are expiring and teams should be notified about.
I admit this is not the most performant solution, but should not present any
challenges for a long time. The alternative approach would be to start changing the
'enabled' attribute on the encryption certs, which adds quite a lot of complexity.