mb664GkE+GmBNuF4o: Fix expiring certs labels
When certs were expiring the tags were not correct, as there was a rounding going on
so certs expiring in 11 hours showed as "EXPIRES IN 1 DAYS" and under 11 hours
showed as "EXPIRES IN 0 DAYS".
Similar with expired certs which showed "EXPIRES IN -2 DAYS" if expired 2 days ago.

This PR adds more granular metrics for hours and minutes and pluralize accordingly.

Also extracted the logic for selecting the right class for the label to the helper.