69ughKnj: Out-of-hours notification
As part of our protection, we want to be notified of any out-of-hours activity which
could introduce a change to the production environment.
This changes adds an additional logic to the production publish event which checks whether
it's out-of-hours (6PM-8AM and weekends) and sends an email if it is.