HUB-744 update ruby version to 2.6.6
First stage of moving to vedorised Docker images is to bring all our
stuff up to the same ruby version. 2.6.6 is a minor point release for
verify self service and is also the same version currently being used
in the verify-frontend so makes sense to bring everything to this
version.