HUB-981: Add support page with email
Added a new static support page to Self-Service instead of redirecting to an external page. The custom page is based on the page that we were redirecting to (https://www.verify.service.gov.uk/support/). The content has been cut down to only include information relevant to RPs or IDPs. A mailto link has been added so RPs and IDPs can find our support email address.

Use the environment variable 'support_email' to get the value for the email. This means the actual email address won't be included in the public repo.

  - Added the `view_support` permission set to `true` - this allows all users to view the support page and only logged in users can see the page
  - Added a test for SupportController

  - Added the `-d` option to the startup script so the user can run the app with docker without providing input on the console
  - Updated README to provide instructions to log in when running locally